Makefile: build and tests for both ARM and AArch64
Detecting compilers, qemu and specifying a way to set default compiler,
target architecture, etc. I'm trying to be as safe as possible, bailing
when some crucial options aren't correct. But others will bail only when
one tries to run the rules (ex. if there are no QEMU binaries).

Clang added for good measure, but it fails to compile due to inline
assembly syntax issues. I'm ok to remove clang support for now to focus
on the general functionality for now.